requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
azure-ai-agents==1.1.0
azure-ai-projects==1.0.0
azure-core==1.35.0
azure-identity==1.23.1
azure-storage-blob==12.26.0
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
cryptography==45.0.6
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.72
langchain-openai==0.3.28
langsmith==0.4.12
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.33.0
msal-extensions==1.3.1
openai==1.99.1
orjson==3.11.1
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.10.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==5.1.1
zope.interface==7.2
zstandard==0.23.0
